requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
aiohttp-socks>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
